+++ CHANGELOG.md
@@ -22,6 +22,11 @@ Find file release
 
 [rug-602]: https://github.com/atomist/rug/issues/602
 
+### Added
+
+-   Added ConvertManifestToPackageJson editor to convert from `manifest.yml``
+    to `package.json`
+
 ## [0.30.0] - 2017-05-12
 
 [0.30.0]: https://github.com/atomist/rug-rugs/compare/0.29.0...0.30.0